Remove Babel macros, support for Node 12.13.0 LTS
